refactor: rename plugin folder
