+++ packages/build-webpack-config/src/webpackPlugins/AssetsManifestPlugin.ts
